fixed merge mistakes
