+++ pkg/addr/isdas.go
@@ -67,10 +67,10 @@ type AS uint64
 // ParseAS parses an AS from a decimal (in the case of the 32bit BGP AS number
 // space) or ipv6-style hex (in the case of SCION-only AS numbers) string.
 func ParseAS(as string) (AS, error) {
-	return ParseASSep(as, ":")
+	return parseAS(as, ":")
 }
 
-func ParseASSep(as string, sep string) (AS, error) {
+func parseAS(as string, sep string) (AS, error) {
 	parts := strings.Split(as, sep)
 	if len(parts) == 1 {
 		// Must be a BGP AS, parse as 32-bit decimal number
