+++ tools/integration/integration.go
@@ -218,8 +218,7 @@ func generateAllSrcDst(hostAddr HostAddr, unique bool) []IAPair {
 
 type HostAddr func(ia addr.IA) *snet.UDPAddr
 
-// CSAddr reads the tester host Addr from the topology for the specified IA.
-// If the address cannot be found, the CS address is returned.
+// CSAddr reads the CS host Addr from the topology for the specified IA.
 var CSAddr HostAddr = func(ia addr.IA) *snet.UDPAddr {
 	if a := loadAddr(ia); a != nil {
 		return a
